# chore(@astrojs/node): use Node.js for testing