--- a/packages/astro/test/i18n-routing.test.js
+++ b/packages/astro/test/i18n-routing.test.js
@@ -2,6 +2,7 @@ import { expect } from 'chai';
 import * as cheerio from 'cheerio';
 import testAdapter from './test-adapter.js';
 import { loadFixture } from './test-utils.js';
+import { describe, it, before, after } from 'node:test';
 
 describe('astro:i18n virtual module', () => {
 	/** @type {import('./test-utils').Fixture} */