--- a/packages/integrations/node/test/errors.test.js
+++ b/packages/integrations/node/test/errors.test.js
@@ -1,6 +1,7 @@
+import * as assert from 'node:assert/strict';
+import { describe, it, before, after } from 'node:test';
 import nodejs from '../dist/index.js';
 import { loadFixture } from './test-utils.js';
-import { expect } from 'chai';
 import * as cheerio from 'cheerio';
 
 describe('Errors', () => {
@@ -13,21 +14,19 @@ describe('Errors', () => {
 		});
 		await fixture.build();
 	});
-	describe('Within the stream', async () => {
-		let devPreview;
+	let devPreview;
 
-		before(async () => {
-			devPreview = await fixture.preview();
-		});
-		after(async () => {
-			await devPreview.stop();
-		});
-		it('when mode is standalone', async () => {
-			const res = await fixture.fetch('/in-stream');
-			const html = await res.text();
-			const $ = cheerio.load(html);
+	before(async () => {
+		devPreview = await fixture.preview();
+	});
+	after(async () => {
+		await devPreview.stop();
+	});
+	it('when mode is standalone', async () => {
+		const res = await fixture.fetch('/in-stream');
+		const html = await res.text();
+		const $ = cheerio.load(html);
 
-			expect($('p').text().trim()).to.equal('Internal server error');
-		});
+		assert.equal($('p').text().trim(), 'Internal server error');
 	});
 });