--- a/packages/integrations/node/test/node-middleware.test.js
+++ b/packages/integrations/node/test/node-middleware.test.js
@@ -1,7 +1,7 @@
 import * as assert from 'node:assert/strict';
 import { describe, it, before, after } from 'node:test';
 import nodejs from '../dist/index.js';
-import { loadFixture } from './test-utils.js';
+import { loadFixture, waitServerListen } from './test-utils.js';
 import * as cheerio from 'cheerio';
 import express from 'express';
 
@@ -32,6 +32,7 @@ describe('behavior from middleware, standalone', () => {
 		const { startServer } = await load();
 		let res = startServer();
 		server = res.server;
+		await waitServerListen(server.server);
 	});
 
 	after(async () => {