… packages/integrations/node/test/url-protocol.test.js
… packages/integrations/node/test/well-known-locations.test.js
… pnpm-lock.yaml