@Princesseuh 2024-01-22
Would love to see if there's any differences in speed!