@ematipico 2024-01-23
This PR

```
Benchmark 1: pnpm test
  Time (mean ± σ):     17.103 s ±  0.678 s    [User: 26.434 s, System: 4.905 s]
  Range (min … max):   16.390 s … 18.307 s    10 runs
```

`main`

```
Benchmark 1: pnpm test
  Time (mean ± σ):      5.786 s ±  0.232 s    [User: 7.962 s, System: 1.420 s]
  Range (min … max):    5.520 s …  6.272 s    10 runs
```