@ematipico 2024-01-23
I am closing this for now because regressions are concerning. Node.js runs tests in parallel, spawning a process for each test. While this decision is understandable, there should be a way to allow to run tests in one single process.

I opened a feature request here: https://github.com/nodejs/node/issues/51548